updates for Chat sample

+++ sample-chat/sample-chat/Services/ChatService.m
@@ -100,7 +100,7 @@ typedef void(^CompletionBlockWithResult)(NSArray *);
     }
 }
 
-- (void)chatDidFailWithError:(int)code{
+- (void)chatDidFailWithError:(NSInteger)code{
     // relogin here
     [[QBChat instance] loginWithUser:self.currentUser];
 }
